refactor(webdav): remove services->commands auto-sync dependency
